pyproject.toml
[project]
name = "search-my-name"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "easyocr>=1.7.2",
    "matplotlib>=3.10.8",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pymupdf>=1.26.7",
    "pytesseract>=0.3.13",
    "torch>=2.5.0",
    "torchvision>=0.20.0",
]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu121" }
torchvision = { index = "pytorch-cu121" }
